+++ nir/src/main/scala/scala/scalanative/nir/serialization/package.scala
@@ -3,14 +3,9 @@ package nir
 
 import java.io.OutputStream
 import java.nio._
+import java.nio.channels.ByteChannel
 
 package object serialization {
-  def serializeText(defns: Seq[Defn], buffer: ByteBuffer): Unit = {
-    val builder = Show.newBuilder
-    builder.defns_(defns)
-    buffer.put(builder.toString.getBytes)
-  }
-
   @inline
   private def withBigEndian[T](buf: ByteBuffer)(body: ByteBuffer => T): T = {
     val o = buf.order()
@@ -19,11 +14,17 @@ package object serialization {
     finally buf.order(o)
   }
 
+  def serializeBinary(defns: Seq[Defn], channel: ByteChannel): Unit = {
+    val writer = new NIRWriter()
+    writer.put(defns)
+    writer.write(channel)
+  } 
   def serializeBinary(defns: Seq[Defn], out: OutputStream): Unit =
-    new BinarySerializer().serialize(defns, out)
+    ???
+    // new BinarySerializer().serialize(defns, out)
 
   def deserializeBinary(buffer: ByteBuffer, bufferName: String): Seq[Defn] =
     withBigEndian(buffer) {
-      new BinaryDeserializer(_, bufferName).deserialize()
+      new BinaryReader(_, bufferName).deserialize()
     }
 }
