+++ nscplugin/src/main/scala-2/scala/scalanative/nscplugin/NirGenFile.scala
@@ -6,6 +6,8 @@ import java.nio.file.{Path, Paths}
 import scala.scalanative.nir.serialization.serializeBinary
 import scala.tools.nsc.Global
 import scala.tools.nsc.io.AbstractFile
+import java.nio.channels.FileChannel
+import java.nio.file.StandardOpenOption
 
 trait NirGenFile[G <: Global with Singleton] { self: NirGenPhase[G] =>
   import global._
@@ -25,9 +27,16 @@ trait NirGenFile[G <: Global with Singleton] { self: NirGenPhase[G] =>
     dir.fileNamed(filename + ".nir")
   }
 
-  def genIRFile(path: AbstractFile, defns: Seq[nir.Defn]): Unit = {
-    val outStream = path.bufferedOutput
-    try serializeBinary(defns, outStream)
-    finally outStream.close()
-  }
+  def genIRFile(path: AbstractFile, defns: Seq[nir.Defn]): Unit =
+    if (path.isVirtual) {
+      val outStream = path.bufferedOutput
+      try serializeBinary(defns, outStream)
+      finally outStream.close()
+    } else {
+      import StandardOpenOption._
+      val channel =
+        FileChannel.open(path.file.toPath(), CREATE, WRITE, TRUNCATE_EXISTING)
+      try serializeBinary(defns, channel)
+      finally channel.close()
+    }
 }
